Add files via upload
This assessment will operate as a "seen exam", operated in an open book test format, so all course materials and use of the internet to search for advice will be allowed. Students will be provided with a worksheet on Monday 8th March, and asked to complete and submit the text exam by 15:00 on Thursday 11th March. A series of five programming tasks will be assigned within the class test. This is designed to replicate a basic programming competency test, as would be expected in a job interview.
Assessment Criteria:

Ability to write a C++ function to a given specification.

Demonstrate an understanding of basic programming concepts, and when are appropriate to use.

Ability to create legible and well presented code.

Ability to follow the brief and completeness of answers